Migrate `/whisper`
