+++ src/CMakeLists.txt
@@ -89,6 +89,8 @@ set(SOURCE_FILES
         controllers/commands/builtin/twitch/RemoveVIP.hpp
         controllers/commands/builtin/twitch/SendReply.cpp
         controllers/commands/builtin/twitch/SendReply.hpp
+        controllers/commands/builtin/twitch/SendWhisper.cpp
+        controllers/commands/builtin/twitch/SendWhisper.hpp
         controllers/commands/builtin/twitch/ShieldMode.cpp
         controllers/commands/builtin/twitch/ShieldMode.hpp
         controllers/commands/builtin/twitch/Shoutout.cpp
