+++ src/controllers/commands/CommandController.cpp
@@ -22,6 +22,7 @@
 #include "controllers/commands/builtin/twitch/RemoveModerator.hpp"
 #include "controllers/commands/builtin/twitch/RemoveVIP.hpp"
 #include "controllers/commands/builtin/twitch/SendReply.hpp"
+#include "controllers/commands/builtin/twitch/SendWhisper.hpp"
 #include "controllers/commands/builtin/twitch/ShieldMode.hpp"
 #include "controllers/commands/builtin/twitch/Shoutout.hpp"
 #include "controllers/commands/builtin/twitch/StartCommercial.hpp"
@@ -77,292 +78,6 @@
 namespace {
 
 using namespace chatterino;
-
-bool areIRCCommandsStillAvailable()
-{
-    // 11th of February 2023, 06:00am UTC
-    const QDateTime migrationTime(QDate(2023, 2, 11), QTime(6, 0), Qt::UTC);
-    auto now = QDateTime::currentDateTimeUtc();
-    return now < migrationTime;
-}
-
-QString useIRCCommand(const QStringList &words)
-{
-    // Reform the original command
-    auto originalCommand = words.join(" ");
-
-    // Replace the / with a . to pass it along to TMI
-    auto newCommand = originalCommand;
-    newCommand.replace(0, 1, ".");
-
-    qCDebug(chatterinoTwitch)
-        << "Forwarding command" << originalCommand << "as" << newCommand;
-
-    return newCommand;
-}
-
-void sendWhisperMessage(const QString &text)
-{
-    // (hemirt) pajlada: "we should not be sending whispers through jtv, but
-    // rather to your own username"
-    auto app = getApp();
-    QString toSend = text.simplified();
-
-    app->twitch->sendMessage("jtv", toSend);
-}
-
-bool appendWhisperMessageWordsLocally(const QStringList &words)
-{
-    auto app = getApp();
-
-    MessageBuilder b;
-
-    b.emplace<TimestampElement>();
-    b.emplace<TextElement>(app->accounts->twitch.getCurrent()->getUserName(),
-                           MessageElementFlag::Text, MessageColor::Text,
-                           FontStyle::ChatMediumBold);
-    b.emplace<TextElement>("->", MessageElementFlag::Text,
-                           getApp()->themes->messages.textColors.system);
-    b.emplace<TextElement>(words[1] + ":", MessageElementFlag::Text,
-                           MessageColor::Text, FontStyle::ChatMediumBold);
-
-    const auto &acc = app->accounts->twitch.getCurrent();
-    const auto &accemotes = *acc->accessEmotes();
-    const auto &bttvemotes = app->twitch->getBttvEmotes();
-    const auto &ffzemotes = app->twitch->getFfzEmotes();
-    auto flags = MessageElementFlags();
-    auto emote = std::optional<EmotePtr>{};
-    for (int i = 2; i < words.length(); i++)
-    {
-        {  // Twitch emote
-            auto it = accemotes.emotes.find({words[i]});
-            if (it != accemotes.emotes.end())
-            {
-                b.emplace<EmoteElement>(it->second,
-                                        MessageElementFlag::TwitchEmote);
-                continue;
-            }
-        }  // Twitch emote
-
-        {  // bttv/ffz emote
-            if ((emote = bttvemotes.emote({words[i]})))
-            {
-                flags = MessageElementFlag::BttvEmote;
-            }
-            else if ((emote = ffzemotes.emote({words[i]})))
-            {
-                flags = MessageElementFlag::FfzEmote;
-            }
-            if (emote)
-            {
-                b.emplace<EmoteElement>(*emote, flags);
-                continue;
-            }
-        }  // bttv/ffz emote
-        {  // emoji/text
-            for (auto &variant : app->emotes->emojis.parse(words[i]))
-            {
-                constexpr const static struct {
-                    void operator()(EmotePtr emote, MessageBuilder &b) const
-                    {
-                        b.emplace<EmoteElement>(emote,
-                                                MessageElementFlag::EmojiAll);
-                    }
-                    void operator()(const QString &string,
-                                    MessageBuilder &b) const
-                    {
-                        LinkParser parser(string);
-                        if (parser.result())
-                        {
-                            b.addLink(*parser.result());
-                        }
-                        else
-                        {
-                            b.emplace<TextElement>(string,
-                                                   MessageElementFlag::Text);
-                        }
-                    }
-                } visitor;
-                boost::apply_visitor(
-                    [&b](auto &&arg) {
-                        visitor(arg, b);
-                    },
-                    variant);
-            }  // emoji/text
-        }
-    }
-
-    b->flags.set(MessageFlag::DoNotTriggerNotification);
-    b->flags.set(MessageFlag::Whisper);
-    auto messagexD = b.release();
-
-    app->twitch->whispersChannel->addMessage(messagexD);
-
-    auto overrideFlags = std::optional<MessageFlags>(messagexD->flags);
-    overrideFlags->set(MessageFlag::DoNotLog);
-
-    if (getSettings()->inlineWhispers &&
-        !(getSettings()->streamerModeSuppressInlineWhispers &&
-          isInStreamerMode()))
-    {
-        app->twitch->forEachChannel(
-            [&messagexD, overrideFlags](ChannelPtr _channel) {
-                _channel->addMessage(messagexD, overrideFlags);
-            });
-    }
-
-    return true;
-}
-
-bool useIrcForWhisperCommand()
-{
-    switch (getSettings()->helixTimegateWhisper.getValue())
-    {
-        case HelixTimegateOverride::Timegate: {
-            if (areIRCCommandsStillAvailable())
-            {
-                return true;
-            }
-
-            // fall through to Helix logic
-        }
-        break;
-
-        case HelixTimegateOverride::AlwaysUseIRC: {
-            return true;
-        }
-        break;
-
-        case HelixTimegateOverride::AlwaysUseHelix: {
-            // do nothing and fall through to Helix logic
-        }
-        break;
-    }
-    return false;
-}
-
-QString runWhisperCommand(const QStringList &words, const ChannelPtr &channel)
-{
-    if (words.size() < 3)
-    {
-        channel->addMessage(
-            makeSystemMessage("Usage: /w <username> <message>"));
-        return "";
-    }
-
-    auto currentUser = getApp()->accounts->twitch.getCurrent();
-    if (currentUser->isAnon())
-    {
-        channel->addMessage(
-            makeSystemMessage("You must be logged in to send a whisper!"));
-        return "";
-    }
-    auto target = words.at(1);
-    stripChannelName(target);
-    auto message = words.mid(2).join(' ');
-    if (channel->isTwitchChannel())
-    {
-        // this covers all twitch channels and twitch-like channels
-        if (useIrcForWhisperCommand())
-        {
-            appendWhisperMessageWordsLocally(words);
-            sendWhisperMessage(words.join(' '));
-            return "";
-        }
-        getHelix()->getUserByName(
-            target,
-            [channel, currentUser, target, message,
-             words](const auto &targetUser) {
-                getHelix()->sendWhisper(
-                    currentUser->getUserId(), targetUser.id, message,
-                    [words] {
-                        appendWhisperMessageWordsLocally(words);
-                    },
-                    [channel, target, targetUser](auto error, auto message) {
-                        using Error = HelixWhisperError;
-
-                        QString errorMessage = "Failed to send whisper - ";
-
-                        switch (error)
-                        {
-                            case Error::NoVerifiedPhone: {
-                                errorMessage +=
-                                    "Due to Twitch restrictions, you are now "
-                                    "required to have a verified phone number "
-                                    "to send whispers. You can add a phone "
-                                    "number in Twitch settings. "
-                                    "https://www.twitch.tv/settings/security";
-                            };
-                            break;
-
-                            case Error::RecipientBlockedUser: {
-                                errorMessage +=
-                                    "The recipient doesn't allow whispers "
-                                    "from strangers or you directly.";
-                            };
-                            break;
-
-                            case Error::WhisperSelf: {
-                                errorMessage += "You cannot whisper yourself.";
-                            };
-                            break;
-
-                            case Error::Forwarded: {
-                                errorMessage += message;
-                            }
-                            break;
-
-                            case Error::Ratelimited: {
-                                errorMessage +=
-                                    "You may only whisper a maximum of 40 "
-                                    "unique recipients per day. Within the "
-                                    "per day limit, you may whisper a "
-                                    "maximum of 3 whispers per second and "
-                                    "a maximum of 100 whispers per minute.";
-                            }
-                            break;
-
-                            case Error::UserMissingScope: {
-                                // TODO(pajlada): Phrase MISSING_REQUIRED_SCOPE
-                                errorMessage += "Missing required scope. "
-                                                "Re-login with your "
-                                                "account and try again.";
-                            }
-                            break;
-
-                            case Error::UserNotAuthorized: {
-                                // TODO(pajlada): Phrase MISSING_PERMISSION
-                                errorMessage += "You don't have permission to "
-                                                "perform that action.";
-                            }
-                            break;
-
-                            case Error::Unknown: {
-                                errorMessage +=
-                                    "An unknown error has occurred.";
-                            }
-                            break;
-                        }
-                        channel->addMessage(makeSystemMessage(errorMessage));
-                    });
-            },
-            [channel] {
-                channel->addMessage(
-                    makeSystemMessage("No user matching that username."));
-            });
-        return "";
-    }
-    // we must be on IRC
-    auto *ircChannel = dynamic_cast<IrcChannel *>(channel.get());
-    if (ircChannel == nullptr)
-    {
-        // give up
-        return "";
-    }
-    auto *server = ircChannel->server();
-    server->sendWhisper(target, message);
-    return "";
-}
 
 using VariableReplacer = std::function<QString(
     const QString &, const ChannelPtr &, const Message *)>;
@@ -827,9 +542,7 @@ void CommandController::initialize(Settings &, Paths &paths)
 
     for (const auto &cmd : TWITCH_WHISPER_COMMANDS)
     {
-        this->registerCommand(cmd, [](const QStringList &words, auto channel) {
-            return runWhisperCommand(words, channel);
-        });
+        this->registerCommand(cmd, &commands::sendWhisper);
     }
 
     this->registerCommand("/vips", &commands::getVIPs);
