+++ src/controllers/commands/builtin/twitch/SendWhisper.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+class QString;
+
+namespace chatterino {
+
+struct CommandContext;
+
+}  // namespace chatterino
+
+namespace chatterino::commands {
+
+QString sendWhisper(const CommandContext &ctx);
+
+}  // namespace chatterino::commands
